+++ packages/core/src/components/ozon-content/nodes/table.node/table-rows.tsx
@@ -5,19 +5,26 @@ import { OzonContentNodeContext } from "../../ozon-content-node-context.interfac
 
 import { Colspecs } from "./colspec/colspec.interface";
 import { Cell } from "./table-cell";
+import { Header } from "./table-header";
 
 export const Rows: FunctionalComponent<{
   context: OzonContentNodeContext;
   colspecs: Colspecs | undefined;
   rows: Element[];
-}> = ({ context, colspecs, rows }) => {
+  head?: boolean;
+}> = ({ context, colspecs, rows, head }) => {
   return (
     <Fragment>
       {rows.map((row) => (
         <tr>
-          {Array.from(row.children).map((cell) => (
-            <Cell cell={cell} colspecs={colspecs} context={context} wijzigactie={parseWijzigactieFromNode(row)} />
-          ))}
+          {Array.from(row.children).map(
+            (cell) =>
+              (head && (
+                <Header cell={cell} colspecs={colspecs} context={context} wijzigactie={parseWijzigactieFromNode(row)} />
+              )) || (
+                <Cell cell={cell} colspecs={colspecs} context={context} wijzigactie={parseWijzigactieFromNode(row)} />
+              ),
+          )}
         </tr>
       ))}
     </Fragment>